fix(evals): close remaining review gaps
