+++ scripts/review/review-eval.mjs
@@ -6,15 +6,7 @@
 // (`run-eval.sh`) invokes it.
 
 import { spawnSync } from "node:child_process";
-import { createHash } from "node:crypto";
-import {
-  appendFileSync,
-  existsSync,
-  lstatSync,
-  readdirSync,
-  readFileSync,
-  realpathSync,
-} from "node:fs";
+import { appendFileSync, existsSync, lstatSync, readFileSync } from "node:fs";
 import path from "node:path";
 import process from "node:process";
 import { parseArgs as parseNodeArgs } from "node:util";
@@ -38,31 +30,28 @@ import {
   completeMatrixProblems,
   freshness,
   frozenDefectProblems,
-  parseInstant,
   readLedger,
   validateLedgerRow,
 } from "./review-eval-ledger.mjs";
+import { revalidateAppendedRows } from "./review-eval-appended.mjs";
+import { planProvenanceProblems } from "./review-eval-plan-evidence.mjs";
 import {
-  baselineEligibility,
   parseLeadingReviewEvalMarkers,
-  leakSuspected,
   renderReport,
   REVIEW_EVAL_OWNERSHIP_LABEL,
   scheduleIssuePayload,
   verdict,
 } from "./review-eval-report.mjs";
+import { runEvidenceProblems } from "./review-eval-run-evidence.mjs";
 import {
   assertAuthorizedFreshnessWorkflow,
   buildPlan,
-  baselinePlanIdentity,
-  cellFingerprint,
   claudeExec,
   comparabilityKey,
   DEFAULT_CALIBRATION_PATH,
   DEFAULT_LEDGER_PATH,
   DEFAULT_RUNS_DIR,
   fileDigest,
-  planCells,
   planStalenessIssueSync,
   resolveKind,
   scorePlan,
@@ -73,6 +62,8 @@ import {
   resolveRowReference,
   revalidateRow,
 } from "./review-eval-result-shape.mjs";
+
+export { planProvenanceProblems };
 
 export const DEFAULT_REVIEW_EVAL_REPO = "mento-protocol/monitoring-monorepo";
 
@@ -512,1017 +503,6 @@ async function modeCheckLedger(options, context) {
   if (!result.ok) process.exitCode = 1;
 }
 
-/** Whether a detail directory holds any scored cell result. */
-function holdsCellResults(dir) {
-  try {
-    return readdirSync(dir).some(
-      (name) => name.startsWith("result-") && name.endsWith(".json"),
-    );
-  } catch {
-    return false;
-  }
-}
-
-/** Reject evidence links and special files before any JSON reader follows them. */
-function nonRegularEvidenceProblems(dir) {
-  if (!existsSync(dir)) return [];
-  const problems = [];
-  for (const file of readdirSync(dir).filter(
-    (name) =>
-      name === "calibration.json" ||
-      name === "plan.json" ||
-      (name.startsWith("result-") && name.endsWith(".json")),
-  )) {
-    try {
-      if (!lstatSync(path.join(dir, file)).isFile()) {
-        problems.push(`${dir}/${file} must be a regular evidence file`);
-      }
-    } catch {
-      problems.push(`${dir}/${file} cannot be inspected as an evidence file`);
-    }
-  }
-  return problems;
-}
-
-/**
- * The row's provenance, checked against the plan that produced it.
- *
- * `revalidateRow` re-derives every number a row states from the cell records
- * beside it, and nothing else. The fields that say which run those records
- * belong to — the contract, the comparability key, the kind, the recorded
- * inputs and the detail directory — it takes from the row. Editing one after
- * the local append therefore survived CI: a changed `comparability_key` on the
- * first full row keeps its no-baseline GREEN verdict, opens a lineage no run
- * produced, and `resolveBaseline` anchors every later run on it.
- *
- * `plan.json` is committed in the same directory as the cell records and is
- * written before the matrix spends anything, so it is the run's own statement
- * of what it was. A directory that holds cell results must hold it: deleting
- * the file cannot buy back the check, exactly as with `calibration.json`. A
- * Only an evidence-free hand-assembled bridge row may omit the plan. It cannot
- * become an automatic anchor or refresh the full-run clock.
- */
-export function planProvenanceProblems({
-  dir,
-  row,
-  contract,
-  baselineRow = null,
-  expectedComparabilityKey = null,
-}) {
-  const problems = [];
-  const checkBaselineStanding = () => {
-    if (baselineRow === null) return;
-    const eligibility = baselineEligibility(baselineRow);
-    if (!eligibility.usable) {
-      problems.push(
-        `explicit baseline ${baselineRow.executed_at} is not eligible: ${eligibility.reason}`,
-      );
-    }
-    if (!(baselineRow.executed_at < row.executed_at)) {
-      problems.push(
-        `explicit baseline ${baselineRow.executed_at} must precede candidate ${row.executed_at}`,
-      );
-    }
-  };
-  if (row.kind === "bridge") checkBaselineStanding();
-  const file = path.join(dir, "plan.json");
-  if (!existsSync(file)) {
-    if (row.kind === "bridge" && !holdsCellResults(dir)) return problems;
-    problems.push(
-      `${dir} carries no plan.json; the row's provenance cannot be checked against the run that produced it`,
-    );
-    return problems;
-  }
-  let plan;
-  try {
-    plan = readJson(file);
-  } catch (error) {
-    problems.push(error instanceof Error ? error.message : String(error));
-    return problems;
-  }
-  /** True when the plan carries the complete explicit-baseline fingerprint. */
-  const validBaselineIdentity = (value) =>
-    value !== null &&
-    typeof value === "object" &&
-    !Array.isArray(value) &&
-    JSON.stringify(Object.keys(value).sort()) ===
-      JSON.stringify(
-        [
-          "comparability_key",
-          "contract_digest",
-          "detail_dir",
-          "executed_at",
-          "row_digest",
-        ].sort(),
-      ) &&
-    parseInstant(value.executed_at) !== null &&
-    /^[0-9a-f]{64}$/.test(value.contract_digest) &&
-    /^[0-9a-f]{64}$/.test(value.comparability_key) &&
-    typeof value.detail_dir === "string" &&
-    value.detail_dir.length > 0 &&
-    /^[0-9a-f]{64}$/.test(value.row_digest);
-  for (const field of [
-    "contract_digest",
-    "comparability_key",
-    "kind",
-    "detail_dir",
-  ]) {
-    // The one kind a plan can legitimately disagree about. No CLI mode plans a
-    // bridge — `--kind` takes `full` and `canary`, and `buildPlan` refuses
-    // anything else — so the model-retirement procedure in
-    // docs/evals/review-skill.md builds the bridge row by hand from the newer
-    // of the two full runs: it copies that run's `row.json`, changes only
-    // `kind`, and validates against that run's own detail directory. Demanding
-    // plan parity on `kind` therefore rejected every bridge row the runbook
-    // can produce, and `--revalidate-appended` is a required step of the
-    // ledger PR. Only `bridge` over a `full` plan is allowed: a bridge over a
-    // canary plan, or any other kind swap, is still an edit, and the other
-    // three fields plus `inputs` are still checked for a bridge row.
-    if (field === "kind" && row.kind === "bridge" && plan.kind === "full") {
-      continue;
-    }
-    if (row[field] !== plan[field]) {
-      problems.push(
-        `row ${field} is ${JSON.stringify(row[field])}; plan.json in ${row.detail_dir} planned ${JSON.stringify(plan[field])}`,
-      );
-    }
-  }
-  if (
-    expectedComparabilityKey !== null &&
-    plan.comparability_key !== expectedComparabilityKey
-  ) {
-    problems.push(
-      `plan.json in ${row.detail_dir} carries comparability_key ${JSON.stringify(plan.comparability_key)}; current frozen inputs derive ${expectedComparabilityKey}`,
-    );
-  }
-  // The inputs are the skill, argv, orchestrator and CLI provenance the scorer
-  // copies out of the plan verbatim, so any difference is an edit.
-  if (JSON.stringify(row.inputs) !== JSON.stringify(plan.inputs)) {
-    problems.push(
-      `row inputs do not match the inputs plan.json in ${row.detail_dir} recorded`,
-    );
-  }
-  if (!["automatic", "explicit"].includes(plan.baseline_selection)) {
-    problems.push(
-      `plan.json in ${row.detail_dir} carries invalid baseline_selection ${JSON.stringify(plan.baseline_selection)}`,
-    );
-  } else if (
-    plan.baseline_selection === "automatic" &&
-    plan.baseline !== null
-  ) {
-    problems.push(
-      `plan.json in ${row.detail_dir} planned automatic baseline selection but carries an explicit baseline identity`,
-    );
-  } else if (
-    plan.baseline_selection === "explicit" &&
-    !validBaselineIdentity(plan.baseline)
-  ) {
-    problems.push(
-      `plan.json in ${row.detail_dir} planned an explicit baseline but carries no valid baseline identity`,
-    );
-  } else if (
-    plan.baseline_selection === "explicit" &&
-    row.vs_baseline === null &&
-    row.kind !== "bridge" &&
-    row.status !== "failed"
-  ) {
-    problems.push(
-      `row has no vs_baseline; plan.json in ${row.detail_dir} planned an explicit baseline`,
-    );
-  } else if (row.vs_baseline !== null && row.kind !== "bridge") {
-    if (row.vs_baseline?.selection !== plan.baseline_selection) {
-      problems.push(
-        `row baseline selection is ${JSON.stringify(row.vs_baseline?.selection)}; plan.json in ${row.detail_dir} planned ${JSON.stringify(plan.baseline_selection)}`,
-      );
-    }
-    if (
-      plan.baseline_selection === "explicit" &&
-      (plan.baseline.executed_at !== row.vs_baseline?.baseline_executed_at ||
-        plan.baseline.comparability_key !==
-          row.vs_baseline?.baseline_comparability_key)
-    ) {
-      problems.push(
-        `row baseline identity does not match the explicit baseline plan.json in ${row.detail_dir} recorded`,
-      );
-    }
-  }
-  if (
-    plan.baseline_selection === "explicit" &&
-    baselineRow !== null &&
-    JSON.stringify(plan.baseline) !==
-      JSON.stringify(baselinePlanIdentity(baselineRow))
-  ) {
-    problems.push(
-      `resolved baseline does not match the explicit baseline identity plan.json in ${row.detail_dir} recorded`,
-    );
-  }
-  if (plan.baseline_selection === "explicit" && row.kind !== "bridge") {
-    checkBaselineStanding();
-  }
-  if (!Array.isArray(plan.cells)) {
-    problems.push(`plan.json in ${row.detail_dir} carries no cells array`);
-    return problems;
-  }
-  if (!contract || !["full", "canary"].includes(plan.kind)) {
-    problems.push(
-      `plan.json in ${row.detail_dir} has no frozen ${plan.kind ?? "unknown"} matrix to validate`,
-    );
-  } else {
-    const expectedCells = planCells({ contract, kind: plan.kind });
-    if (JSON.stringify(plan.cells) !== JSON.stringify(expectedCells)) {
-      problems.push(
-        `plan.json in ${row.detail_dir} cells do not match the frozen ${plan.kind} matrix`,
-      );
-    }
-  }
-  for (const [name, condition] of Object.entries(row.conditions ?? {})) {
-    const cells = plan.cells.filter((cell) => cell.condition === name);
-    if (cells.length === 0) {
-      problems.push(
-        `row condition ${name} has no matching cell in plan.json in ${row.detail_dir}`,
-      );
-      continue;
-    }
-    for (const field of ["model", "effort", "finder"]) {
-      const planned = [...new Set(cells.map((cell) => cell[field]))];
-      if (planned.length !== 1) {
-        problems.push(
-          `plan.json in ${row.detail_dir} records ${planned.length} ${field} values for condition ${name}`,
-        );
-      } else if (condition?.[field] !== planned[0]) {
-        problems.push(
-          `row condition ${name}.${field} is ${JSON.stringify(condition?.[field])}; plan.json recorded ${JSON.stringify(planned[0])}`,
-        );
-      }
-    }
-  }
-  return problems;
-}
-
-/** Validate required fields that every scored result must retain. */
-function resultEvidenceProblems({ dir, row }) {
-  if (row.status === "failed" || !existsSync(dir)) return [];
-  const regularityProblems = nonRegularEvidenceProblems(dir);
-  if (regularityProblems.length > 0) return regularityProblems;
-  const problems = [];
-  let expectedFingerprint = null;
-  try {
-    expectedFingerprint = cellFingerprint({
-      plan: readJson(path.join(dir, "plan.json")),
-    });
-  } catch {
-    // planProvenanceProblems reports a missing or unreadable plan.
-  }
-  if (holdsCellResults(dir)) {
-    const scoringUsd = row?.scoring_usd;
-    if (
-      !Object.hasOwn(row ?? {}, "scoring_usd") ||
-      typeof scoringUsd !== "number" ||
-      !Number.isFinite(scoringUsd) ||
-      scoringUsd < 0
-    ) {
-      problems.push(
-        "row.scoring_usd must be a nonnegative finite number for scored evidence",
-      );
-    }
-  }
-  let resultRecordsLeak = false;
-  for (const resultFile of readdirSync(dir).filter(
-    (name) => name.startsWith("result-") && name.endsWith(".json"),
-  )) {
-    try {
-      const record = readJson(path.join(dir, resultFile));
-      if (
-        expectedFingerprint !== null &&
-        JSON.stringify(record?.fingerprint) !==
-          JSON.stringify(expectedFingerprint)
-      ) {
-        problems.push(
-          `${dir}/${resultFile} fingerprint does not match the plan execution inputs`,
-        );
-      }
-      const scoringUsd = record?.scoring_usd;
-      if (
-        !Object.hasOwn(record ?? {}, "scoring_usd") ||
-        typeof scoringUsd !== "number" ||
-        !Number.isFinite(scoringUsd) ||
-        scoringUsd < 0
-      ) {
-        problems.push(
-          `${dir}/${resultFile} scoring_usd must be a nonnegative finite number`,
-        );
-      }
-      const leak = record?.leak;
-      const hard = leak?.hard;
-      const suspected = leak?.suspected;
-      if (
-        leak === null ||
-        typeof leak !== "object" ||
-        Array.isArray(leak) ||
-        !Array.isArray(hard) ||
-        typeof suspected !== "boolean"
-      ) {
-        problems.push(
-          `${dir}/${resultFile} leak must carry a hard array and suspected boolean`,
-        );
-      } else if (suspected !== hard.length > 0) {
-        problems.push(
-          `${dir}/${resultFile} leak.suspected must equal whether leak.hard is non-empty`,
-        );
-      }
-      if (suspected === true || (Array.isArray(hard) && hard.length > 0)) {
-        resultRecordsLeak = true;
-      }
-    } catch {
-      // The normal evidence checks report unreadable result records.
-    }
-  }
-  const calibrationFile = path.join(dir, "calibration.json");
-  if (existsSync(calibrationFile)) {
-    try {
-      const calibration = readJson(calibrationFile);
-      if (
-        expectedFingerprint !== null &&
-        JSON.stringify(calibration?.fingerprint) !==
-          JSON.stringify(expectedFingerprint)
-      ) {
-        problems.push(
-          `${dir}/calibration.json fingerprint does not match the plan execution inputs`,
-        );
-      }
-      const scoringUsd = calibration?.scoring_usd;
-      if (
-        !Object.hasOwn(calibration ?? {}, "scoring_usd") ||
-        typeof scoringUsd !== "number" ||
-        !Number.isFinite(scoringUsd) ||
-        scoringUsd < 0
-      ) {
-        problems.push(
-          `${dir}/calibration.json scoring_usd must be a nonnegative finite number`,
-        );
-      }
-    } catch {
-      // The calibration evidence checks report an unreadable record.
-    }
-  }
-  if (resultRecordsLeak && !leakSuspected(row)) {
-    problems.push(
-      `${dir} carries a result with leak.suspected true, but the row notes omit leak suspected`,
-    );
-  }
-  if (resultRecordsLeak && row.verdict !== "AMBER") {
-    problems.push(
-      `${dir} carries a result with leak.suspected true, but the row verdict is ${row.verdict} instead of AMBER`,
-    );
-  }
-  return problems;
-}
-
-/** Evidence files required for every row that claims scored results. */
-function runEvidenceProblems({ dir, row, contract }) {
-  const regularityProblems = nonRegularEvidenceProblems(dir);
-  if (regularityProblems.length > 0) return regularityProblems;
-  if (row.status === "failed") {
-    return readdirSync(dir)
-      .filter(
-        (name) =>
-          name === "calibration.json" ||
-          (name.startsWith("result-") && name.endsWith(".json")),
-      )
-      .map(
-        (name) =>
-          `${dir} failed row retains scoring artifact ${name}; failed traces must clear scored evidence`,
-      );
-  }
-  const problems = resultEvidenceProblems({ dir, row });
-  // A bridge intentionally reuses the source full run's plan and scored
-  // records. It owes no new evidence, but it must retain any leak flag those
-  // records carry because that flag makes the reused score unusable.
-  if (row.kind === "bridge") return problems;
-  if (!holdsCellResults(dir)) {
-    problems.push(`${dir} carries no scored result-*.json files`);
-  }
-  const planFile = path.join(dir, "plan.json");
-  if (existsSync(planFile)) {
-    try {
-      const plan = readJson(planFile);
-      if (Array.isArray(plan.cells)) {
-        const rowConditions = new Set(Object.keys(row.conditions ?? {}));
-        const planConditions = new Set(
-          plan.cells.map((cell) => cell.condition),
-        );
-        const plannedResults = new Map(
-          plan.cells.map((cell) => [
-            `result-${cell.pr}-${cell.condition}-${cell.draw}.json`,
-            cell,
-          ]),
-        );
-        const resultFiles = readdirSync(dir).filter(
-          (name) => name.startsWith("result-") && name.endsWith(".json"),
-        );
-        try {
-          const completedCellIds = readJson(
-            path.join(dir, "calibration.json"),
-          )?.completed_cell_ids;
-          if (
-            !Array.isArray(completedCellIds) ||
-            completedCellIds.some(
-              (cellId) => typeof cellId !== "string" || cellId.length === 0,
-            )
-          ) {
-            problems.push(
-              `${dir}/calibration.json completed_cell_ids must be an array of non-empty strings`,
-            );
-          } else {
-            const plannedCellIds = new Set(
-              plan.cells.map((cell) => cell.cell_id),
-            );
-            const completedSet = new Set(completedCellIds);
-            if (completedSet.size !== completedCellIds.length) {
-              problems.push(
-                `${dir}/calibration.json completed_cell_ids must not contain duplicates`,
-              );
-            }
-            for (const cellId of completedSet) {
-              if (!plannedCellIds.has(cellId)) {
-                problems.push(
-                  `${dir}/calibration.json records unplanned completed cell ${cellId}`,
-                );
-              }
-            }
-            const resultCellIds = resultFiles
-              .map((resultFile) => plannedResults.get(resultFile)?.cell_id)
-              .filter(Boolean);
-            const sorted = (values) => [...values].sort();
-            if (
-              JSON.stringify(sorted(completedSet)) !==
-              JSON.stringify(sorted(new Set(resultCellIds)))
-            ) {
-              problems.push(
-                `${dir} result files do not match calibration.json completed_cell_ids`,
-              );
-            }
-            const expectedStatus =
-              completedSet.size === plannedCellIds.size
-                ? "complete"
-                : "partial";
-            if (row.status !== expectedStatus) {
-              problems.push(
-                `${dir} records ${row.status} status but calibration.json proves a ${expectedStatus} matrix`,
-              );
-            }
-          }
-        } catch {
-          // The calibration evidence checks report an unreadable record.
-        }
-        for (const resultFile of resultFiles) {
-          const cell = plannedResults.get(resultFile);
-          if (!cell) {
-            problems.push(`${dir} carries unplanned result file ${resultFile}`);
-            continue;
-          }
-          let record;
-          try {
-            record = readJson(path.join(dir, resultFile));
-          } catch (error) {
-            problems.push(
-              error instanceof Error ? error.message : String(error),
-            );
-            continue;
-          }
-          for (const field of ["cell_id", "pr", "condition", "draw"]) {
-            if (record?.[field] !== cell[field]) {
-              problems.push(
-                `${dir}/${resultFile} ${field} is ${JSON.stringify(record?.[field])}; plan.json recorded ${JSON.stringify(cell[field])}`,
-              );
-            }
-          }
-          const fixture = contract.fixtures.find(
-            (candidate) => candidate.pr === cell.pr,
-          );
-          if (!Array.isArray(record?.matched_ids)) {
-            problems.push(`${dir}/${resultFile} matched_ids must be an array`);
-          } else if (fixture) {
-            const allowedIds = new Set(fixture.scorable_ids.map(String));
-            for (const id of record.matched_ids) {
-              if (
-                !(
-                  typeof id === "string" ||
-                  (typeof id === "number" && Number.isSafeInteger(id))
-                )
-              ) {
-                problems.push(
-                  `${dir}/${resultFile} matched_ids contains non-scalar ${JSON.stringify(id)}`,
-                );
-              } else if (!allowedIds.has(String(id))) {
-                problems.push(
-                  `${dir}/${resultFile} matched_ids contains ${JSON.stringify(id)}, which fixture PR ${cell.pr} does not score`,
-                );
-              }
-            }
-          }
-          const claims = record?.claims;
-          if (!Array.isArray(claims)) {
-            problems.push(`${dir}/${resultFile} claims must be an array`);
-          } else {
-            for (const claim of claims) {
-              if (typeof claim !== "string" || claim.trim().length === 0) {
-                problems.push(
-                  `${dir}/${resultFile} claims must contain only non-empty strings`,
-                );
-                break;
-              }
-            }
-          }
-          for (const field of [
-            "claims",
-            "novelWrong",
-            "novelReal",
-            "novelVague",
-            "restatedKnown",
-            "alreadyMatched",
-          ]) {
-            const value = Number(record?.novel?.[field]);
-            if (!Number.isSafeInteger(value) || value < 0) {
-              problems.push(
-                `${dir}/${resultFile} novel.${field} must be a nonnegative safe integer`,
-              );
-            }
-          }
-          if (
-            Array.isArray(claims) &&
-            record?.novel?.claims !== claims.length
-          ) {
-            problems.push(
-              `${dir}/${resultFile} novel.claims must equal the claims array length`,
-            );
-          }
-          if (
-            Array.isArray(record?.matched_ids) &&
-            record?.novel?.alreadyMatched !== record.matched_ids.length
-          ) {
-            problems.push(
-              `${dir}/${resultFile} novel.alreadyMatched must equal the matched_ids array length`,
-            );
-          }
-          const verdicts = record?.novel?.verdicts;
-          if (
-            verdicts === null ||
-            typeof verdicts !== "object" ||
-            Array.isArray(verdicts)
-          ) {
-            problems.push(
-              `${dir}/${resultFile} novel.verdicts must be an object`,
-            );
-          } else if (Array.isArray(claims)) {
-            const expectedKeys = claims.map((_claim, index) =>
-              String(index + 1),
-            );
-            const actualKeys = Object.keys(verdicts).sort(
-              (left, right) => Number(left) - Number(right),
-            );
-            if (JSON.stringify(actualKeys) !== JSON.stringify(expectedKeys)) {
-              problems.push(
-                `${dir}/${resultFile} novel.verdicts must carry one numbered verdict per claim`,
-              );
-            }
-            const counts = {
-              real: 0,
-              wrong: 0,
-              vague: 0,
-              known: 0,
-            };
-            let validVerdicts = true;
-            for (const verdict of Object.values(verdicts)) {
-              if (
-                verdict === null ||
-                typeof verdict !== "object" ||
-                Array.isArray(verdict) ||
-                !Object.hasOwn(counts, verdict.class)
-              ) {
-                validVerdicts = false;
-                break;
-              }
-              counts[verdict.class] += 1;
-            }
-            if (!validVerdicts) {
-              problems.push(
-                `${dir}/${resultFile} novel.verdicts carries an invalid class`,
-              );
-            } else {
-              for (const [field, count] of [
-                ["novelReal", counts.real],
-                ["novelWrong", counts.wrong],
-                ["novelVague", counts.vague],
-                ["restatedKnown", counts.known],
-              ]) {
-                if (record?.novel?.[field] !== count) {
-                  problems.push(
-                    `${dir}/${resultFile} novel.${field} does not match novel.verdicts`,
-                  );
-                }
-              }
-            }
-          }
-          for (const field of ["usd", "seconds"]) {
-            const value = record?.[field];
-            if (
-              typeof value !== "number" ||
-              !Number.isFinite(value) ||
-              value < 0
-            ) {
-              problems.push(
-                `${dir}/${resultFile} ${field} must be a nonnegative finite number`,
-              );
-            }
-          }
-          if (!rowConditions.has(cell.condition)) {
-            problems.push(
-              `${dir}/${resultFile} records condition ${cell.condition}, but the row omits it`,
-            );
-          }
-        }
-        if (row.status === "complete") {
-          for (const condition of planConditions) {
-            if (!rowConditions.has(condition)) {
-              problems.push(
-                `${dir} planned condition ${condition}, but the complete row omits it`,
-              );
-            }
-          }
-          for (const condition of rowConditions) {
-            if (!planConditions.has(condition)) {
-              problems.push(
-                `${dir} complete row condition ${condition} has no planned cells`,
-              );
-            }
-          }
-          for (const cell of plan.cells) {
-            const resultFile = `result-${cell.pr}-${cell.condition}-${cell.draw}.json`;
-            if (!resultFiles.includes(resultFile)) {
-              problems.push(
-                `${dir} carries no ${resultFile} for planned cell ${cell.cell_id ?? "unknown"}`,
-              );
-            }
-          }
-        } else {
-          const cells = plan.cells.filter((cell) =>
-            rowConditions.has(cell.condition),
-          );
-          for (const condition of rowConditions) {
-            const resultCells = cells.filter(
-              (cell) =>
-                cell.condition === condition &&
-                resultFiles.includes(
-                  `result-${cell.pr}-${cell.condition}-${cell.draw}.json`,
-                ),
-            );
-            if (resultCells.length === 0) {
-              problems.push(
-                `${dir} carries no scored result for row condition ${condition}`,
-              );
-              continue;
-            }
-            const representedPrs = new Set(resultCells.map((cell) => cell.pr));
-            const perDefect = row.conditions?.[condition]?.per_defect ?? {};
-            for (const pr of representedPrs) {
-              const fixture = contract.fixtures.find(
-                (candidate) => candidate.pr === pr,
-              );
-              const missingIds = (fixture?.scorable_ids ?? [])
-                .map(String)
-                .filter((id) => !Object.hasOwn(perDefect, id));
-              if (missingIds.length > 0) {
-                problems.push(
-                  `${dir} row condition ${condition} omits frozen defect ${missingIds.join(", ")} for scored result PR ${pr}`,
-                );
-              }
-            }
-          }
-        }
-      }
-    } catch {
-      // planProvenanceProblems reports the unreadable plan with its full error.
-    }
-  }
-  if (!existsSync(path.join(dir, "calibration.json"))) {
-    problems.push(`${dir} carries no calibration.json`);
-  }
-  return problems;
-}
-
-/** Canonical JSON text for evidence fingerprints. */
-function canonicalJson(value) {
-  if (Array.isArray(value)) {
-    return `[${value.map(canonicalJson).join(",")}]`;
-  }
-  if (value !== null && typeof value === "object") {
-    return `{${Object.keys(value)
-      .sort()
-      .map((key) => `${JSON.stringify(key)}:${canonicalJson(value[key])}`)
-      .join(",")}}`;
-  }
-  return JSON.stringify(value);
-}
-
-/** Digest one run's exact canonical scored evidence, independent of layout. */
-function runEvidenceDigest(dir) {
-  if (!existsSync(dir)) return null;
-  if (nonRegularEvidenceProblems(dir).length > 0) return null;
-  const files = readdirSync(dir)
-    .filter(
-      (name) =>
-        name === "calibration.json" ||
-        (name.startsWith("result-") && name.endsWith(".json")),
-    )
-    .sort();
-  if (files.length === 0) return null;
-  const hash = createHash("sha256");
-  const update = (value) => {
-    const bytes = Buffer.from(value);
-    const length = Buffer.alloc(8);
-    length.writeBigUInt64BE(BigInt(bytes.length));
-    hash.update(length);
-    hash.update(bytes);
-  };
-  try {
-    for (const file of files) {
-      update(file);
-      update(canonicalJson(readJson(path.join(dir, file))));
-    }
-  } catch {
-    return null;
-  }
-  return hash.digest("hex");
-}
-
-/**
- * Recompute every row this branch appends, from the detail the same branch
- * commits. `checkLedger` proves a ledger PR is schema-valid, covers the frozen
- * ids and adds rows without editing older ones — and nothing more. Editing a
- * committed row's `verdict`, its counters or a `per_defect` bit before opening
- * the PR left all three checks green while the committed report no longer
- * matched its own cell and calibration evidence, and `--validate` ran only on
- * the operator's machine before the commit.
- *
- * No model is called: `revalidateRow` reads the committed `result-*.json` and
- * `calibration.json` files. The job that runs this holds no model credential.
- */
-function revalidateAppendedRows({ options, context, result, base }) {
-  const problems = [];
-  const rows = result.rows;
-  // Which rows are new is the base comparison's answer. An unresolvable base
-  // ref leaves that unanswerable — recomputing the whole history would judge
-  // rows of retired contracts, and checking nothing is a guard that silently
-  // no-ops. A base that resolves but carries no ledger is the bootstrap case:
-  // every row on this branch is appended, zero rows included.
-  if (!base.rows && !base.resolved) {
-    result.problems.push(
-      `--revalidate-appended cannot tell which rows are new: ${base.reason}`,
-    );
-    return { ok: false, checked: null, unpaired: null };
-  }
-  const appended = base.rows ? rows.slice(base.rows.length) : rows;
-  const calibrationFile = path.resolve(
-    context.repoRoot,
-    options.calibrationPath,
-  );
-  const calibrationSet = existsSync(calibrationFile)
-    ? readJson(calibrationFile)
-    : null;
-  const expectedComparabilityKey = comparabilityKey({
-    contract: context.contract,
-    contractDigest: context.contractDigest,
-    calibrationDigest: fileDigest(calibrationFile),
-  });
-  const repoRoot = realpathSync(context.repoRoot);
-  const detailDirectoryIdentity = (detailDir, label) => {
-    const resolved = path.resolve(repoRoot, detailDir);
-    const insideRepo = (target) => {
-      const relative = path.relative(repoRoot, target);
-      return (
-        relative === "" ||
-        (relative !== ".." &&
-          !relative.startsWith(`..${path.sep}`) &&
-          !path.isAbsolute(relative))
-      );
-    };
-    if (!insideRepo(resolved)) {
-      problems.push(`${label} names detail_dir ${detailDir} outside the repo`);
-      return null;
-    }
-    const canonical = existsSync(resolved) ? realpathSync(resolved) : resolved;
-    if (!insideRepo(canonical)) {
-      problems.push(`${label} names detail_dir ${detailDir} outside the repo`);
-      return null;
-    }
-    if (canonical === repoRoot) {
-      problems.push(`${label} names the repository root as detail_dir`);
-      return null;
-    }
-    return canonical;
-  };
-  const usedDetailDirs = new Set();
-  const usedEvidence = new Map();
-  const baseDetailPaths = new Map();
-  for (const row of base.rows ?? []) {
-    if (typeof row.detail_dir !== "string") continue;
-    const identity = detailDirectoryIdentity(
-      row.detail_dir,
-      `base row ${row.executed_at}`,
-    );
-    if (identity !== null) {
-      usedDetailDirs.add(identity);
-      baseDetailPaths.set(
-        path
-          .relative(repoRoot, path.resolve(repoRoot, row.detail_dir))
-          .split(path.sep)
-          .join("/"),
-        row.executed_at,
-      );
-      baseDetailPaths.set(
-        path.relative(repoRoot, identity).split(path.sep).join("/"),
-        row.executed_at,
-      );
-      problems.push(
-        ...nonRegularEvidenceProblems(identity).map(
-          (problem) => `base row ${row.executed_at}: ${problem}`,
-        ),
-      );
-      if (row.kind !== "bridge" && row.status !== "failed") {
-        const digest = runEvidenceDigest(identity);
-        if (digest !== null) {
-          usedEvidence.set(digest, `base row ${row.executed_at}`);
-        }
-      }
-    }
-  }
-  if (base.rows && base.base) {
-    const changed = spawnSync(
-      "git",
-      ["diff", "--name-only", "-z", "--no-renames", base.base, "--"],
-      { cwd: repoRoot },
-    );
-    if (changed.status !== 0) {
-      problems.push(
-        `could not compare base evidence directories with ${base.base}`,
-      );
-    } else {
-      for (const changedPath of String(changed.stdout)
-        .split("\0")
-        .filter(Boolean)) {
-        for (const [detailPath, executedAt] of baseDetailPaths) {
-          if (
-            changedPath === detailPath ||
-            changedPath.startsWith(`${detailPath}/`) ||
-            detailPath.startsWith(`${changedPath}/`)
-          ) {
-            problems.push(
-              `base row ${executedAt} evidence changed at ${changedPath}`,
-            );
-            break;
-          }
-        }
-      }
-    }
-  }
-  let unpaired = 0;
-  for (const row of appended) {
-    const label = `appended row ${row.executed_at}`;
-    if (!row.detail_dir) {
-      problems.push(`${label} carries no detail_dir to recompute from`);
-      continue;
-    }
-    const dir = detailDirectoryIdentity(row.detail_dir, label);
-    if (dir === null) continue;
-    if (row.kind !== "bridge" && usedDetailDirs.has(dir)) {
-      problems.push(
-        `${label} reuses detail_dir ${row.detail_dir} from an earlier ledger row`,
-      );
-      continue;
-    }
-    usedDetailDirs.add(dir);
-    if (!existsSync(dir)) {
-      problems.push(
-        `${label} names detail_dir ${row.detail_dir}, which this branch does not commit`,
-      );
-      continue;
-    }
-    problems.push(
-      ...nonRegularEvidenceProblems(dir).map(
-        (problem) => `${label}: ${problem}`,
-      ),
-    );
-    if (row.kind !== "bridge" && row.status !== "failed") {
-      const digest = runEvidenceDigest(dir);
-      const previous = digest === null ? null : usedEvidence.get(digest);
-      if (previous) {
-        problems.push(
-          `${label} reuses scored result and calibration evidence from ${previous}`,
-        );
-      } else if (digest !== null) {
-        usedEvidence.set(digest, label);
-      }
-    }
-    // The recorded pairing is rechecked against the row it actually names, not
-    // against whatever anchor this ledger would resolve on its own — those are
-    // two different pairs of runs, and comparing the wrong one reports
-    // differences that are not errors. A candidate row scored against an
-    // installed row whose own ledger PR has not merged yet names a row this
-    // branch does not carry; that pairing cannot be rechecked here, so the row
-    // is recomputed against its own bits and detail alone and counted as
-    // unpaired rather than failed.
-    const skipsBaselinePairing = row.status === "failed";
-    const recordedAt = row.vs_baseline?.baseline_executed_at ?? null;
-    const baselineRow =
-      recordedAt === null
-        ? null
-        : (rows.find((candidate) => candidate.executed_at === recordedAt) ??
-          null);
-    const missingRecordedBaseline = recordedAt !== null && baselineRow === null;
-    const baselineSelection = row.vs_baseline?.selection ?? null;
-    const candidateMissingBaseline =
-      missingRecordedBaseline &&
-      row.inputs?.dirty === true &&
-      row.inputs?.skill_ref !== "installed" &&
-      baselineSelection === "explicit";
-    if (
-      !skipsBaselinePairing &&
-      missingRecordedBaseline &&
-      !candidateMissingBaseline
-    ) {
-      problems.push(
-        `${label}: baseline ${recordedAt} is not in the ledger; only a dirty --skill-ref candidate row with explicit selection may waive a missing baseline`,
-      );
-    }
-    const baselineMissing = candidateMissingBaseline;
-    if (baselineMissing) unpaired += 1;
-    if (
-      !skipsBaselinePairing &&
-      recordedAt !== null &&
-      baselineSelection === null
-    ) {
-      problems.push(
-        `${label}: vs_baseline.selection must record automatic or explicit baseline selection`,
-      );
-    }
-    const baselineIsExplicit =
-      row.kind === "bridge" || baselineSelection === "explicit";
-    if (!skipsBaselinePairing && !baselineMissing && !baselineIsExplicit) {
-      const resolvedBaseline = resolveBaseline({ rows, row });
-      const sameBaseline =
-        resolvedBaseline === baselineRow ||
-        (resolvedBaseline !== null &&
-          baselineRow !== null &&
-          resolvedBaseline.executed_at === baselineRow.executed_at &&
-          resolvedBaseline.contract_digest === baselineRow.contract_digest &&
-          resolvedBaseline.detail_dir === baselineRow.detail_dir);
-      if (
-        !sameBaseline &&
-        !(resolvedBaseline === null && baselineRow === null)
-      ) {
-        problems.push(
-          `${label}: recorded baseline ${recordedAt ?? "none"} does not match the append-order baseline ${resolvedBaseline?.executed_at ?? "none"}`,
-        );
-      }
-    }
-    const check = revalidateRow({
-      contract: context.contract,
-      row,
-      repoRoot: context.repoRoot,
-      detailDir: dir,
-      ledgerRows: baselineMissing || skipsBaselinePairing ? [] : rows,
-      baselineRow: skipsBaselinePairing ? null : baselineRow,
-      baselineIsExplicit,
-      // Half the verdict is the pairing: a regression and a promotion are both
-      // read out of the flip counts against the anchor. Recomputing an unpaired
-      // row without one turns a recorded RED or PROMOTE into GREEN and fails
-      // the ledger PR for it, over a row the runbook says only counts toward
-      // `unpaired_baselines`. The detail checks above are unaffected — they
-      // read the row's own bits and its own cell records — so they still run.
-      baselineMissing,
-      calibrationSet,
-    });
-    problems.push(...check.problems.map((problem) => `${label}: ${problem}`));
-    problems.push(
-      ...planProvenanceProblems({
-        dir,
-        row,
-        contract: context.contract,
-        baselineRow,
-        expectedComparabilityKey,
-      }).map((problem) => `${label}: ${problem}`),
-    );
-    problems.push(
-      ...runEvidenceProblems({ dir, row, contract: context.contract }).map(
-        (problem) => `${label}: ${problem}`,
-      ),
-    );
-  }
-  result.problems.push(...problems);
-  return { ok: problems.length === 0, checked: appended.length, unpaired };
-}
-
 async function modePlan(options, context) {
   const rows = readLedger(path.resolve(context.repoRoot, options.ledgerPath));
   const baselineRow = resolveRowReference({
@@ -1661,6 +641,11 @@ async function modeValidate(options, context) {
   const validationLedgerRows = options.append
     ? [...ledgerRows, row]
     : ledgerRows;
+  const baselineRow = resolveRowReference({
+    reference: options.against,
+    rows: ledgerRows,
+    repoRoot: context.repoRoot,
+  });
   const revalidated = revalidateRow({
     contract: context.contract,
     row,
@@ -1671,11 +656,7 @@ async function modeValidate(options, context) {
     ledgerRows: validationLedgerRows,
     // Name the same baseline `--score --against` used, or the row's own
     // verdict is rechecked against a baseline it was never scored on.
-    baselineRow: resolveRowReference({
-      reference: options.against,
-      rows: ledgerRows,
-      repoRoot: context.repoRoot,
-    }),
+    baselineRow,
     // The frozen pairs the recorded calibration outcomes must be about, so a
     // detail file cannot relabel what the judge was expected to answer. A row
     // validated outside a checkout that carries the set still has its recorded
@@ -1688,6 +669,45 @@ async function modeValidate(options, context) {
   if (row.contract_digest !== context.contractDigest) {
     problems.push(
       `row contract_digest ${row.contract_digest?.slice(0, 8)} is not the current contract`,
+    );
+  }
+  let detailPathExists = false;
+  let hasDetailDirectory = false;
+  try {
+    if (typeof revalidated.detail_dir === "string") {
+      detailPathExists = true;
+      hasDetailDirectory = lstatSync(revalidated.detail_dir).isDirectory();
+    }
+  } catch {
+    detailPathExists = false;
+    // The problem below names the missing or uninspectable directory.
+  }
+  if (!hasDetailDirectory && (options.append || detailPathExists)) {
+    problems.push(
+      `${options.append ? "--validate --append" : "--validate"} requires an existing non-symlink detail directory; found ${JSON.stringify(revalidated.detail_dir)}`,
+    );
+  }
+  if (hasDetailDirectory) {
+    const expectedComparabilityKey = existsSync(calibrationFile)
+      ? comparabilityKey({
+          contract: context.contract,
+          contractDigest: context.contractDigest,
+          calibrationDigest: fileDigest(calibrationFile),
+        })
+      : null;
+    problems.push(
+      ...planProvenanceProblems({
+        dir: revalidated.detail_dir,
+        row,
+        contract: context.contract,
+        baselineRow,
+        expectedComparabilityKey,
+      }),
+      ...runEvidenceProblems({
+        dir: revalidated.detail_dir,
+        row,
+        contract: context.contract,
+      }),
     );
   }
   // `revalidateRow` recomputes the conditions the row lists, and over the ids
