+++ scripts/review/review-eval.test.mjs
@@ -48,10 +48,10 @@ import {
   cellReuseDecision,
   claudeArgv,
   comparabilityKey,
-  fileDigest,
   leakSignals,
   loginsInFixtureTree,
-  ORCHESTRATOR_FILE,
+  ORCHESTRATOR_FILES,
+  orchestratorSourceDigest,
   planCells,
   planStalenessIssueSync,
   resolveDetailDir,
@@ -60,6 +60,7 @@ import {
   skillDigest,
   SCRUBBED_ENV_VARS,
   scrubbedEnv,
+  treatmentIdentity,
 } from "./review-eval-run.mjs";
 import {
   buildVsBaseline,
@@ -75,6 +76,17 @@ import {
 
 const repoRoot = fileURLToPath(new URL("../..", import.meta.url));
 const scriptPath = fileURLToPath(new URL("./review-eval.mjs", import.meta.url));
+const validationModuleLineLimits = new Map([
+  ["review-eval.mjs", 900],
+  ["review-eval-run.mjs", 100],
+  ["review-eval-run-plan.mjs", 600],
+  ["review-eval-run-execution.mjs", 600],
+  ["review-eval-run-cell.mjs", 600],
+  ["review-eval-run-score.mjs", 600],
+  ["review-eval-plan-evidence.mjs", 600],
+  ["review-eval-run-evidence.mjs", 600],
+  ["review-eval-appended.mjs", 600],
+]);
 const contractRelative = "docs/evals/review-skill-fixtures.json";
 const ledgerRelative = "docs/evals/review-skill-ledger.jsonl";
 const { contract, digest: contractDigest } = loadContract(
@@ -86,6 +98,78 @@ const planEnv = {
   REVIEW_EVAL_HOST: "test-host",
   REVIEW_EVAL_SKILL_DIR: path.join(repoRoot, "scripts/review/prompts"),
 };
+
+const runEvalSourcePaths = new Map([
+  ["wrapper", path.join(repoRoot, "scripts/review/run-eval.sh")],
+  ["lifecycle", path.join(repoRoot, "scripts/review/run-eval-lifecycle.sh")],
+  ["runtime", path.join(repoRoot, "scripts/review/run-eval-runtime.sh")],
+]);
+
+function runEvalSource(owner) {
+  const sourcePath = runEvalSourcePaths.get(owner);
+  assert.ok(sourcePath, `unknown run-eval source owner: ${owner}`);
+  return readFileSync(sourcePath, "utf8");
+}
+
+function runEvalSourceSet() {
+  return [...runEvalSourcePaths]
+    .map(([owner]) => `# source-owner: ${owner}\n${runEvalSource(owner)}`)
+    .join("\n");
+}
+
+function reconstructLegacyOrchestrator() {
+  let wrapper = runEvalSource("wrapper");
+  const helpers = `${runEvalSource("lifecycle")}\n${runEvalSource("runtime")}`;
+  for (const id of [
+    "lifecycle-setup",
+    "lifecycle-verify",
+    "lifecycle-support",
+    "cell-runtime",
+  ]) {
+    const payload = helpers.match(
+      new RegExp(
+        `# RUN-EVAL-ORIGINAL-BEGIN ${id}\\n([\\s\\S]*?)# RUN-EVAL-ORIGINAL-END ${id}\\n`,
+      ),
+    )?.[1];
+    assert.notEqual(payload, undefined, `missing original payload ${id}`);
+    const block = new RegExp(
+      `# RUN-EVAL-EXTRACT-BEGIN ${id}\\n[\\s\\S]*?# RUN-EVAL-EXTRACT-END ${id}\\n`,
+    );
+    assert.match(wrapper, block);
+    wrapper = wrapper.replace(block, () => payload);
+  }
+  return wrapper;
+}
+
+test("run-eval shell sources retain split headroom", () => {
+  for (const owner of runEvalSourcePaths.keys()) {
+    const source = runEvalSource(owner);
+    const lines = source.endsWith("\n")
+      ? source.slice(0, -1).split("\n").length
+      : source.split("\n").length;
+    assert.ok(lines < 600, `${owner} run-eval source has ${lines} lines`);
+  }
+});
+
+test("the shell split reconstructs the cached-cell orchestrator bytes", () => {
+  assert.equal(
+    createHash("sha256").update(reconstructLegacyOrchestrator()).digest("hex"),
+    "5cdfbd0e709af2d68c193d484b724706b339ab0562d14b283f5fc38eebe9ae49",
+  );
+});
+
+test("review-eval validation modules retain split headroom", () => {
+  for (const [name, limit] of validationModuleLineLimits) {
+    const source = readFileSync(
+      fileURLToPath(new URL(`./${name}`, import.meta.url)),
+      "utf8",
+    );
+    const lines = source.endsWith("\n")
+      ? source.slice(0, -1).split("\n").length
+      : source.split("\n").length;
+    assert.ok(lines < limit, `${name} has ${lines} lines; limit is ${limit}`);
+  }
+});
 
 /** A temporary repository root holding only what the CLI reads. */
 function makeRoot() {
@@ -260,6 +344,7 @@ function writeRowEvidence(root, row) {
     cells,
   };
   const fingerprint = cellFingerprint({ plan });
+  const treatment = treatmentIdentity({ plan });
   const completedCellIds = [];
   for (const [name, condition] of Object.entries(row.conditions)) {
     const idsByPr = new Map();
@@ -293,6 +378,7 @@ function writeRowEvidence(root, row) {
           JSON.stringify({
             cell_id: `pr-${pr}-${name}-draw${draw}`,
             fingerprint,
+            treatment,
             pr,
             condition: name,
             draw,
@@ -334,6 +420,7 @@ function writeRowEvidence(root, row) {
     path.join(detail, "calibration.json"),
     JSON.stringify({
       fingerprint,
+      treatment,
       completed_cell_ids: completedCellIds,
       outcomes: calibration.records.map((record) => ({
         record_id: record.record_id,
@@ -643,9 +730,9 @@ test("comparabilityKey moves with the contract, the prompts, and the scorer", ()
       calibrationDigest: "2".repeat(64),
     }),
   );
-  // `run-eval.sh` fixes the contestant's tools, turn limit, skill staging and
-  // finder truncation: it shapes the transcript every recorded number comes
-  // from, so an edit to it re-anchors the series the way a prompt edit does.
+  // The wrapper and both sourced helpers fix the contestant's tools, turn
+  // limit, skill staging and finder truncation. An edit to any of them
+  // re-anchors the series the way a prompt edit does.
   assert.notEqual(
     base,
     comparabilityKey({
@@ -659,9 +746,40 @@ test("comparabilityKey moves with the contract, the prompts, and the scorer", ()
     comparabilityKey({
       contract,
       contractDigest,
-      orchestratorDigest: fileDigest(ORCHESTRATOR_FILE),
+      orchestratorDigest: orchestratorSourceDigest(),
     }),
   );
+});
+
+test("orchestratorSourceDigest binds the wrapper and both helpers", () => {
+  const expected =
+    "77bba1e0af554775f19429d48ea6470a3574b05e6b3ed95a1b3e73e8bf3a2807";
+  assert.equal(orchestratorSourceDigest(), expected);
+  assert.deepEqual(
+    ORCHESTRATOR_FILES.map((file) => path.basename(file)),
+    ["run-eval.sh", "run-eval-lifecycle.sh", "run-eval-runtime.sh"],
+  );
+
+  for (const changed of ORCHESTRATOR_FILES) {
+    const dir = mkdtempSync(path.join(tmpdir(), "review-eval-orchestrator-"));
+    try {
+      const copies = ORCHESTRATOR_FILES.map((file) => {
+        const copy = path.join(dir, path.basename(file));
+        cpSync(file, copy);
+        return copy;
+      });
+      const changedCopy = copies.find(
+        (file) => path.basename(file) === path.basename(changed),
+      );
+      writeFileSync(
+        changedCopy,
+        `${readFileSync(changedCopy, "utf8")}# drift\n`,
+      );
+      assert.notEqual(orchestratorSourceDigest({ files: copies }), expected);
+    } finally {
+      rmSync(dir, { recursive: true, force: true });
+    }
+  }
 });
 
 test("resolveKind picks full only when the last full run is past cadence", () => {
@@ -802,17 +920,54 @@ test("a cached cell from another skill or contract is never reused", () => {
     "claude_cli",
     "codex_cli",
     "contract_digest",
-    "dirty",
     "finder_argv_digest",
     "kind",
     "orchestrator_digest",
     "skill_digest",
-    "skill_ref",
   ]);
+  const candidatePlan = {
+    ...plan,
+    inputs: {
+      ...plan.inputs,
+      skill_ref: "/tmp/review-candidate",
+      dirty: true,
+    },
+  };
+  assert.deepEqual(cellFingerprint({ plan: candidatePlan }), fingerprint);
+  assert.deepEqual(treatmentIdentity({ plan }), {
+    skill_ref: "installed",
+    dirty: false,
+  });
+  assert.deepEqual(treatmentIdentity({ plan: candidatePlan }), {
+    skill_ref: "/tmp/review-candidate",
+    dirty: true,
+  });
 
   const decide = (result) =>
     cellReuseDecision({ plan, resultPath: "unused", result });
   assert.equal(decide({ ok: true, fingerprint }).reuse, true);
+  const legacyOrchestratorFingerprint = {
+    ...fingerprint,
+    orchestrator_digest:
+      "5cdfbd0e709af2d68c193d484b724706b339ab0562d14b283f5fc38eebe9ae49",
+  };
+  const transitioned = decide({
+    ok: true,
+    fingerprint: legacyOrchestratorFingerprint,
+  });
+  assert.equal(transitioned.reuse, true);
+  assert.match(transitioned.reason, /recorded pure orchestrator split/);
+  assert.equal(
+    cellReuseDecision({
+      plan: {
+        ...plan,
+        inputs: { ...plan.inputs, orchestrator_digest: "4".repeat(64) },
+      },
+      resultPath: "unused",
+      result: { ok: true, fingerprint: legacyOrchestratorFingerprint },
+    }).reuse,
+    false,
+  );
   assert.match(decide({ ok: true }).reason, /carries no fingerprint/);
   assert.match(
     decide({ ok: true, fingerprint: { ...fingerprint, skill_digest: "0" } })
@@ -843,10 +998,8 @@ test("a cached cell from another skill or contract is never reused", () => {
   for (const field of [
     "claude_cli",
     "codex_cli",
-    "dirty",
     "finder_argv_digest",
     "orchestrator_digest",
-    "skill_ref",
   ]) {
     assert.match(
       decide({ ok: true, fingerprint: { ...fingerprint, [field]: "moved" } })
@@ -1287,10 +1440,7 @@ test("run-eval.sh refuses a detail_dir before it can delete the checkout", () =>
   // `json_field` prints `String(doc[key])`: a missing key arrives as the word
   // "undefined" and a JSON null as "null". An empty value is the dangerous one
   // — it makes `rm -rf "$REPO/$detail"` target the checkout itself.
-  const shell = readFileSync(
-    path.join(repoRoot, "scripts/review/run-eval.sh"),
-    "utf8",
-  );
+  const shell = runEvalSourceSet();
   const guard = shell.match(
     /\nrequire_safe_detail\(\) \{\n[\s\S]*?\n\}\n/,
   )?.[0];
@@ -1334,10 +1484,7 @@ test("run-eval.sh exits non-zero when no PR carries the row", () => {
   // --pr. Exiting zero while only the publish commands were printed reports a
   // clean run while the next scheduled run refuses to start against a ledger
   // with uncommitted changes.
-  const shell = readFileSync(
-    path.join(repoRoot, "scripts/review/run-eval.sh"),
-    "utf8",
-  );
+  const shell = runEvalSourceSet();
   const guard = shell.match(
     /\nif \[\[ \$PUBLISHED -ne 1 \]\]; then\n[\s\S]*?\nfi\nexit 0\n/,
   )?.[0];
@@ -1370,10 +1517,7 @@ test("run-eval.sh seeds a fresh run directory from the cells it superseded", () 
   // execution its own. The cells in the old one are paid for, so they are
   // copied across — and re-checked one by one against this run's fingerprint,
   // exactly as cells found in place are.
-  const shell = readFileSync(
-    path.join(repoRoot, "scripts/review/run-eval.sh"),
-    "utf8",
-  );
+  const shell = runEvalSourceSet();
   const block = shell.match(
     /\nRESUME_FROM="\$\(json_field "\$PLAN_OUT" resume_from\)"\n[\s\S]*?\nfi\n/,
   )?.[0];
@@ -1442,10 +1586,7 @@ test("run-eval.sh publishes the appended row when the report fails", () => {
   // — a same-sitting `--against` file under /tmp that is gone by now is enough
   // — printed no PR and no recovery commands, and the next scheduled run then
   // refuses to start against the dirty ledger it left behind.
-  const shell = readFileSync(
-    path.join(repoRoot, "scripts/review/run-eval.sh"),
-    "utf8",
-  );
+  const shell = runEvalSourceSet();
   const block = shell.match(
     /\nREPORT="\$RUN_DIR\/report\.md"\n[\s\S]*?\nlog "verdict \$VERDICT"\n/,
   )?.[0];
@@ -1492,10 +1633,7 @@ test("run-eval.sh keeps the stderr of a bounded command it had to fail", () => {
   // A finder, a contestant or the scorer says why it exited on stderr, and
   // every failure path logs only an exit status. Discarding stderr left the
   // one line that explains the failure unreachable.
-  const shell = readFileSync(
-    path.join(repoRoot, "scripts/review/run-eval.sh"),
-    "utf8",
-  );
+  const shell = runEvalSourceSet();
   const parts = ["run_bounded", "log_stderr_tail"].map((name) => {
     const source = shell.match(
       new RegExp(`\\n${name}\\(\\) \\{\\n[\\s\\S]*?\\n\\}\\n`),
@@ -1541,10 +1679,7 @@ test("run-eval.sh refuses a skill staging that did not land", () => {
   // The preamble is taken in a command substitution, so a failed `cp -R` and a
   // preamble missing the instructions both used to render as a plausible,
   // empty treatment that the cell was scored on.
-  const shell = readFileSync(
-    path.join(repoRoot, "scripts/review/run-eval.sh"),
-    "utf8",
-  );
+  const shell = runEvalSourceSet();
   const parts = ["purge_skill", "skill_body_head", "stage_skill"].map(
     (name) => {
       const source = shell.match(
@@ -1627,10 +1762,7 @@ test("run-eval.sh inserts finder output into the handoff prompt verbatim", () =>
   // $' and $1 in a *string* replacement, so a review containing one of those
   // would rewrite the prompt the treatment under test receives. This runs the
   // shell script's own node program to prove the replacement stays literal.
-  const shell = readFileSync(
-    path.join(repoRoot, "scripts/review/run-eval.sh"),
-    "utf8",
-  );
+  const shell = runEvalSourceSet();
   const program = shell.match(
     /REVIEW_EVAL_OTHER="\$other_review" node -e '\n([\s\S]*?)\n {4}' "\$SPEC/,
   )?.[1];
@@ -1651,7 +1783,7 @@ test("run-eval.sh inserts finder output into the handoff prompt verbatim", () =>
   }
 });
 
-test("--validate refuses a bad row and appends a good one", () => {
+test("--validate refuses bad or evidence-free rows and appends a good one", () => {
   const root = makeRoot();
   try {
     const rowPath = path.join(root, "row.json");
@@ -1670,6 +1802,17 @@ test("--validate refuses a bad row and appends a good one", () => {
 
     row.conditions.pipeline.p1.matched -= 1;
     writeFileSync(rowPath, JSON.stringify(row, null, 2));
+    const missing = cli(["--validate", rowPath, "--append", "--json"], {
+      root,
+    });
+    assert.equal(missing.status, 1);
+    assert.match(
+      JSON.parse(missing.stdout).problems.join(" | "),
+      /requires an existing non-symlink detail directory/,
+    );
+    assert.equal(readLedger(path.join(root, ledgerRelative)).length, 0);
+
+    writeRowEvidence(root, row);
     const good = cli(["--validate", rowPath, "--append", "--json"], { root });
     assert.equal(good.status, 0, good.stderr);
     assert.equal(JSON.parse(good.stdout).appended, true);
@@ -1714,6 +1857,33 @@ test("--validate --detail-dir recomputes from a run outside --root", () => {
   } finally {
     rmSync(root, { recursive: true, force: true });
     rmSync(detail, { recursive: true, force: true });
+  }
+});
+
+test("--validate rejects a symlinked detail directory without append", () => {
+  const root = makeRoot();
+  try {
+    const row = makeRow({
+      matchedIds: scorableIdsFor([1990, 1999]),
+      fullMatrix: true,
+    });
+    const detail = writeRowEvidence(root, row);
+    const linkedDetail = path.join(root, "linked-detail");
+    symlinkSync(detail, linkedDetail, "dir");
+    const rowPath = path.join(root, "row.json");
+    writeFileSync(rowPath, JSON.stringify(row, null, 2));
+
+    const result = cli(
+      ["--validate", rowPath, "--detail-dir", linkedDetail, "--json"],
+      { root },
+    );
+    assert.equal(result.status, 1);
+    assert.match(
+      JSON.parse(result.stdout).problems.join(" | "),
+      /--validate requires an existing non-symlink detail directory/,
+    );
+  } finally {
+    rmSync(root, { recursive: true, force: true });
   }
 });
 
@@ -1853,6 +2023,16 @@ test("--validate --against rechecks the verdict on the named baseline", () => {
       fullMatrix: true,
     });
     row.verdict = verdict({ contract, row, baselineRow }).verdict;
+    row.vs_baseline = buildVsBaseline({
+      row,
+      baselineRow,
+      selection: "explicit",
+    });
+    const detail = writeRowEvidence(root, row);
+    const planPath = path.join(detail, "plan.json");
+    const plan = JSON.parse(readFileSync(planPath, "utf8"));
+    plan.baseline = baselinePlanIdentity(baselineRow);
+    writeFileSync(planPath, JSON.stringify(plan));
     const rowPath = path.join(root, "row.json");
     writeFileSync(rowPath, JSON.stringify(row, null, 2));
     const result = cli(
@@ -3002,6 +3182,53 @@ test("scorePlan freezes truth before the calibration pass", async () => {
   }
 });
 
+test("scorePlan verifies the exact truth bytes it snapshots", async () => {
+  const root = makeRoot();
+  try {
+    const plan = buildPlan({
+      contract,
+      contractDigest,
+      kind: "canary",
+      repoRoot: root,
+      outDir: path.join(root, "run"),
+      env: planEnv,
+    });
+    for (const cell of plan.cells) {
+      writeCell(plan, cell, { output: "nothing looks wrong here", root });
+    }
+    const fixture = contract.fixtures.find(
+      (candidate) => candidate.pr === plan.cells[0].pr,
+    );
+    const truthFile = path.join(root, fixture.truth_file);
+    writeFileSync(truthFile, `${readFileSync(truthFile, "utf8")} `);
+    let modelCalls = 0;
+    await assert.rejects(
+      scorePlan({
+        plan,
+        contract,
+        contractDigest,
+        repoRoot: root,
+        planDir: plan.plan_dir,
+        exec: async () => {
+          modelCalls += 1;
+          throw new Error("model must not run");
+        },
+        runGit: stubGit().runGit,
+        calibrationSet: JSON.parse(
+          readFileSync(
+            path.join(root, "docs/evals/review-skill-judge-calibration.json"),
+            "utf8",
+          ),
+        ),
+      }),
+      /truth .* changed after frozen-input verification/,
+    );
+    assert.equal(modelCalls, 0);
+  } finally {
+    rmSync(root, { recursive: true, force: true });
+  }
+});
+
 test("the login exclusions are snapshotted before the tool-bearing judge", async () => {
   // The novelty judge runs with `Bash` inside the fixture. Fixture content that
   // prompt-injects it into writing a reviewer login into a tracked file would,
@@ -3529,6 +3756,12 @@ test("scorePlan reports a partial matrix and refuses an empty one", async () => 
       ),
     );
     assert.deepEqual(scoredResult.fingerprint, canonicalFingerprint);
+    assert.deepEqual(scoredResult.treatment, treatmentIdentity({ plan }));
+    const scoredCalibration = JSON.parse(
+      readFileSync(path.join(plan.plan_dir, "calibration.json"), "utf8"),
+    );
+    assert.deepEqual(scoredCalibration.fingerprint, canonicalFingerprint);
+    assert.deepEqual(scoredCalibration.treatment, treatmentIdentity({ plan }));
     // A canary that did not finish never ranks, not even as a pass.
     assert.equal(scored.row.verdict, "INCOMPLETE");
   } finally {
@@ -4039,10 +4272,7 @@ test("a cell inherits no path back to the source checkout", () => {
   try {
     const fixture = path.join(dir, "fixture");
     mkdirSync(fixture, { recursive: true });
-    const script = readFileSync(
-      path.join(repoRoot, "scripts/review/run-eval.sh"),
-      "utf8",
-    );
+    const script = runEvalSourceSet();
     const cellEnv = script.match(
       /\nCELL_ENV=\(env\n[\s\S]*?PATH="\$CELL_PATH"\)\n/,
     )?.[0];
@@ -4097,10 +4327,7 @@ test("a cell inherits no path back to the source checkout", () => {
 });
 
 test("the orchestrator keeps its resume cache outside the spec worktree", () => {
-  const script = readFileSync(
-    path.join(repoRoot, "scripts/review/run-eval.sh"),
-    "utf8",
-  );
+  const script = runEvalSourceSet();
   // The spec worktree is a temporary directory the EXIT trap removes, so a
   // plan directory inside it takes every completed cell down with it.
   assert.match(script, /RUN_DIR="\$REPO\/\$DETAIL_DIR"/);
@@ -4114,10 +4341,7 @@ test("the spec worktree is not created where a cell can list it", () => {
   // no PR number, reviewer login or withheld SHA for `leakSignals()` to catch.
   // Permissions cannot help — a cell runs as the same user — so it goes under
   // the git directory, which nothing hands a cell a path to.
-  const script = readFileSync(
-    path.join(repoRoot, "scripts/review/run-eval.sh"),
-    "utf8",
-  );
+  const script = runEvalSourceSet();
   assert.match(script, /SPEC="\$\(mktemp -d "\$LOCK_ROOT\/review-eval-spec/);
   assert.equal(/mktemp -d "\$TMPROOT\/review-eval-spec/.test(script), false);
   // `$TMPDIR` is swept by the OS and the git directory is not, so a removal
@@ -4173,16 +4397,13 @@ test("a failed run publishes no partial scoring artifacts", () => {
   }
 });
 
-/** Extract one shell function from run-eval.sh, or fail the test. */
+/** Extract one shell function from the run-eval source that owns it. */
 function shellFunction(name) {
-  const script = readFileSync(
-    path.join(repoRoot, "scripts/review/run-eval.sh"),
-    "utf8",
-  );
+  const script = runEvalSourceSet();
   const source = script.match(
     new RegExp(`\\n${name}\\(\\) \\{\\n[\\s\\S]*?\\n\\}\\n`),
   )?.[0];
-  assert.ok(source, `${name} was not found in run-eval.sh`);
+  assert.ok(source, `${name} was not found in the run-eval source set`);
   return source;
 }
 
@@ -4486,38 +4707,34 @@ test("one review eval at a time may hold the shared run state", async () => {
 });
 
 test("the run lock is taken before anything touches the fixtures", () => {
-  const script = readFileSync(
-    path.join(repoRoot, "scripts/review/run-eval.sh"),
-    "utf8",
-  );
+  const wrapper = runEvalSource("wrapper");
+  const lifecycle = runEvalSource("lifecycle");
   // Taken before the spec worktree, the plan and the matrix, and released by
   // the EXIT trap so an interrupted run does not wedge the next one.
   assert.ok(
-    script.indexOf("\nacquire_run_lock\n") <
-      script.indexOf("# --- the spec worktree"),
+    wrapper.indexOf("# RUN-EVAL-EXTRACT-BEGIN lifecycle-setup") <
+      wrapper.indexOf("# --- the spec worktree"),
     "the run lock is taken after the spec worktree is added",
   );
+  assert.match(lifecycle, /\nacquire_run_lock\n/);
   assert.match(
-    script,
+    lifecycle,
     /for lock_dir in \$\{LOCK_DIRS\[@\]\+"\$\{LOCK_DIRS\[@\]\}"\}; do\n\s*rm -rf "\$lock_dir"/,
   );
   // The checkout half is anchored to the git directory, which no option moves,
   // and it is taken first so the ordering between the two roots is fixed.
   assert.match(
-    script,
+    lifecycle,
     /LOCK_ROOT="\$\(git -C "\$REPO" rev-parse --absolute-git-dir/,
   );
   assert.ok(
-    script.indexOf('acquire_one_lock "$LOCK_ROOT"') <
-      script.indexOf('acquire_one_lock "$CACHE_DIR"'),
+    lifecycle.indexOf('acquire_one_lock "$LOCK_ROOT"') <
+      lifecycle.indexOf('acquire_one_lock "$CACHE_DIR"'),
   );
 });
 
 test("the orchestrator rejects a cell whose finder or report failed", () => {
-  const script = readFileSync(
-    path.join(repoRoot, "scripts/review/run-eval.sh"),
-    "utf8",
-  );
+  const script = runEvalSourceSet();
   // A finder that writes a partial report and then exits non-zero — a session
   // limit, a killed process — produces output that is not empty and is not a
   // review. Cached, it would score forever as a finder that missed those
@@ -4534,10 +4751,7 @@ test("the orchestrator rejects a cell whose finder or report failed", () => {
 });
 
 test("the orchestrator carries one baseline through plan, score, validate and report", () => {
-  const script = readFileSync(
-    path.join(repoRoot, "scripts/review/run-eval.sh"),
-    "utf8",
-  );
+  const script = runEvalSourceSet();
   // The candidate procedure runs the installed skill and the candidate in one
   // sitting. Without a baseline argument the candidate resolves the ledger's
   // stored anchor instead, which is the model drift the procedure exists to
@@ -4560,10 +4774,7 @@ test("the orchestrator carries one baseline through plan, score, validate and re
 });
 
 test("the orchestrator rejects an ineligible baseline before paid work", () => {
-  const script = readFileSync(
-    path.join(repoRoot, "scripts/review/run-eval.sh"),
-    "utf8",
-  );
+  const script = runEvalSourceSet();
   assert.equal(script.match(/baselineEligibility\(row\)/g)?.length, 2);
   assert.ok(
     script.indexOf("baselineEligibility(row)") <
@@ -4613,10 +4824,7 @@ test("the orchestrator rejects an ineligible baseline before paid work", () => {
 });
 
 test("the installed baseline survives the checkout the candidate needs", () => {
-  const script = readFileSync(
-    path.join(repoRoot, "scripts/review/run-eval.sh"),
-    "utf8",
-  );
+  const script = runEvalSourceSet();
   const doc = readFileSync(
     path.join(repoRoot, "docs/evals/review-skill.md"),
     "utf8",
@@ -4643,10 +4851,7 @@ test("the installed baseline survives the checkout the candidate needs", () => {
 });
 
 test("the ledger branch names one run, not one day", () => {
-  const script = readFileSync(
-    path.join(repoRoot, "scripts/review/run-eval.sh"),
-    "utf8",
-  );
+  const script = runEvalSourceSet();
   // Two runs finishing on the same UTC day — the installed and candidate pair
   // the runbook asks for — collided on a date-only branch at `git checkout -b`
   // or at the push, after the paid run and the ledger append were both done.
@@ -4656,10 +4861,7 @@ test("the ledger branch names one run, not one day", () => {
 });
 
 test("the orchestrator snapshots the skill once and refuses a mid-run edit", () => {
-  const script = readFileSync(
-    path.join(repoRoot, "scripts/review/run-eval.sh"),
-    "utf8",
-  );
+  const script = runEvalSourceSet();
   // The plan records one skill digest for the whole matrix and every cached
   // cell's fingerprint carries it, so cells must all stage the same bytes.
   assert.match(script, /SKILL_DIR="\$SKILL_SNAPSHOT"/);
@@ -4672,10 +4874,7 @@ test("the orchestrator snapshots the skill once and refuses a mid-run edit", () 
 });
 
 test("a failed run records its failure, publishes it, or exits non-zero", () => {
-  const script = readFileSync(
-    path.join(repoRoot, "scripts/review/run-eval.sh"),
-    "utf8",
-  );
+  const script = runEvalSourceSet();
   // A run that leaves no trace is indistinguishable from one that never ran,
   // and the freshness guard cannot tell those apart either.
   assert.match(
@@ -4698,10 +4897,7 @@ test("a failed run records its failure, publishes it, or exits non-zero", () => 
 });
 
 test("the run deadline bounds the cell subprocesses and the judge pass", () => {
-  const script = readFileSync(
-    path.join(repoRoot, "scripts/review/run-eval.sh"),
-    "utf8",
-  );
+  const script = runEvalSourceSet();
   // Checked only between cells, the deadline bounded nothing: a finder or a
   // contestant that stalls never returns to the check, and the judge pass ran
   // outside the budget entirely. Every one of the three is now started through
@@ -4739,10 +4935,7 @@ test("the deadline terminates the whole subprocess tree, not only its parent", (
   // Run the committed function itself rather than grepping it: whether a
   // grandchild dies is a property of the process group, which no pattern in the
   // source can assert.
-  const script = readFileSync(
-    path.join(repoRoot, "scripts/review/run-eval.sh"),
-    "utf8",
-  );
+  const script = runEvalSourceSet();
   const definition = script.match(/^run_bounded\(\) \{\n[\s\S]*?^\}$/m);
   assert.ok(definition, "run_bounded is not defined at column zero");
   const fastDefinition = definition[0].replace("sleep 10", "sleep 1");
@@ -4801,15 +4994,49 @@ test("the deadline terminates the whole subprocess tree, not only its parent", (
 });
 
 test("the orchestrator refuses to run bytes the row would not record", () => {
-  const script = readFileSync(
-    path.join(repoRoot, "scripts/review/run-eval.sh"),
-    "utf8",
-  );
+  const wrapper = runEvalSource("wrapper");
+  const lifecycle = runEvalSource("lifecycle");
   // `orchestrator_digest` is taken from the spec worktree, which is where the
   // harness reads every other input. Running an edited copy against a clean
-  // spec would record the spec's bytes for a matrix this file actually shaped.
-  assert.match(script, /ORCHESTRATOR="\$SPEC\/scripts\/review\/run-eval\.sh"/);
-  assert.match(script, /cmp -s "\$\{BASH_SOURCE\[0\]\}" "\$ORCHESTRATOR"/);
+  // spec would record the spec's bytes for a matrix these files actually
+  // shaped. Resolve both helpers from the wrapper path and compare all three.
+  assert.match(
+    wrapper,
+    /RUN_EVAL_SCRIPT_DIR="\$\(cd "\$\(dirname "\$\{BASH_SOURCE\[0\]\}"\)" && pwd\)"/,
+  );
+  for (const helper of ["run-eval-lifecycle.sh", "run-eval-runtime.sh"]) {
+    assert.match(
+      wrapper,
+      new RegExp(
+        `source "\\$RUN_EVAL_SCRIPT_DIR/${helper.replace(".", "\\.")}"`,
+      ),
+    );
+  }
+
+  const verify = lifecycle.match(
+    /\n {2}verify\)\n([\s\S]*?)\n {4};;\n {2}support\)/,
+  )?.[1];
+  assert.ok(verify, "the lifecycle verify stage is missing");
+  for (const source of [
+    "run-eval.sh",
+    "run-eval-lifecycle.sh",
+    "run-eval-runtime.sh",
+  ]) {
+    assert.match(
+      verify,
+      new RegExp(`\\n      ${source.replace(".", "\\.")}\\n`),
+    );
+  }
+  assert.match(
+    verify,
+    /cmp -s "\$RUN_EVAL_RUNNING_SOURCE" "\$RUN_EVAL_SPEC_SOURCE"/,
+  );
+  assert.doesNotMatch(verify, /\$\{BASH_SOURCE\[0\]\}/);
+  assert.match(
+    lifecycle,
+    /: <<'RUN_EVAL_ORIGINAL_LIFECYCLE_VERIFY'[\s\S]*# RUN-EVAL-ORIGINAL-BEGIN lifecycle-verify/,
+  );
+  assert.doesNotMatch(lifecycle, /\n {2}original-verify\)/);
 });
 
 test("the cell reader emits nothing when the plan carries a forged field", () => {
@@ -4817,10 +5044,7 @@ test("the cell reader emits nothing when the plan carries a forged field", () =>
   // substitution and cannot see the writer die. Emitting rows as it goes would
   // run every cell before the offending one and then score that truncated
   // matrix as merely partial, which is what the tab check exists to prevent.
-  const shell = readFileSync(
-    path.join(repoRoot, "scripts/review/run-eval.sh"),
-    "utf8",
-  );
+  const shell = runEvalSourceSet();
   const program = shell.match(
     /cell_rows\(\) \{\n[\s\S]*?node -e '\n([\s\S]*?)\n {2}' "\$PLAN_JSON"/,
   )?.[1];
@@ -5543,10 +5767,7 @@ test("--revalidate-appended treats a base without a ledger as all-appended", () 
 });
 
 test("a scheduled run refuses a checkout that is not at origin/main", () => {
-  const script = readFileSync(
-    path.join(repoRoot, "scripts/review/run-eval.sh"),
-    "utf8",
-  );
+  const script = runEvalSourceSet();
   // The spec worktree pins the contract, but the ledger, the baseline it
   // resolves and the PR commands come from the checkout launchd fires in. A
   // feature branch there would score against the wrong anchor and offer to
@@ -6545,6 +6766,22 @@ test("--revalidate-appended binds candidate status to execution evidence", () =>
     const plan = JSON.parse(readFileSync(planFile, "utf8"));
     plan.inputs = row.inputs;
     writeFileSync(planFile, JSON.stringify(plan));
+    const rowPath = path.join(detail, "row.json");
+    writeFileSync(rowPath, JSON.stringify(row));
+    const validated = cli(["--validate", rowPath, "--json"], { root });
+    assert.equal(validated.status, 1);
+    const validationProblems = JSON.parse(validated.stdout).problems.join(
+      " | ",
+    );
+    assert.match(
+      validationProblems,
+      /result-.* treatment does not match the plan execution inputs/,
+    );
+    assert.match(
+      validationProblems,
+      /calibration\.json treatment does not match the plan execution inputs/,
+    );
+
     writeFileSync(path.join(root, ledgerRelative), `${JSON.stringify(row)}\n`);
 
     const checked = cli(
@@ -6561,11 +6798,11 @@ test("--revalidate-appended binds candidate status to execution evidence", () =>
     const problems = JSON.parse(checked.stdout).problems.join(" | ");
     assert.match(
       problems,
-      /result-.* fingerprint does not match the plan execution inputs/,
+      /result-.* treatment does not match the plan execution inputs/,
     );
     assert.match(
       problems,
-      /calibration\.json fingerprint does not match the plan execution inputs/,
+      /calibration\.json treatment does not match the plan execution inputs/,
     );
   } finally {
     rmSync(root, { recursive: true, force: true });
@@ -6699,6 +6936,18 @@ test("--revalidate-appended accepts an explicit bridge pairing", () => {
       force: true,
     });
     mkdirSync(path.join(root, newer.detail_dir), { recursive: true });
+    const missingBridgeEvidence = cli(flags, { root });
+    assert.equal(missingBridgeEvidence.status, 1);
+    const missingBridgeProblems = JSON.parse(
+      missingBridgeEvidence.stdout,
+    ).problems.join(" | ");
+    assert.match(missingBridgeProblems, /carries no plan\.json/);
+    assert.match(
+      missingBridgeProblems,
+      /carries no scored result-\*\.json files/,
+    );
+    writeRowEvidence(root, newer);
+
     const retiringVerdict = retiring.verdict;
     retiring.notes = "leak suspected: test fixture";
     retiring.verdict = verdict({ contract, row: retiring }).verdict;
@@ -6830,10 +7079,7 @@ test("run-eval.sh resets a fixture to the commit the contract pins", () => {
   // committed its own edits — or a prompt-injected commit out of the diff under
   // review — makes that commit the fixture for every later cell of the PR, for
   // the novelty judge and for the pre-judge login snapshot.
-  const shell = readFileSync(
-    path.join(repoRoot, "scripts/review/run-eval.sh"),
-    "utf8",
-  );
+  const shell = runEvalSourceSet();
   const reset = shell.match(/\nreset_fixture\(\) \{\n[\s\S]*?\n\}\n/)?.[0];
   assert.ok(reset, "reset_fixture was not found in run-eval.sh");
   const dir = mkdtempSync(path.join(tmpdir(), "review-eval-reset-"));
@@ -6886,10 +7132,7 @@ test("run-eval.sh resets a fixture to the commit the contract pins", () => {
 });
 
 test("every cell resets its fixture through reset_fixture", () => {
-  const shell = readFileSync(
-    path.join(repoRoot, "scripts/review/run-eval.sh"),
-    "utf8",
-  );
+  const shell = runEvalSourceSet();
   // The pinned head travels beside the path, and no bare reset survives: an
   // argument-free one is the whole defect.
   assert.match(shell, /reset_fixture "\$fixture" "\$fixture_head"/);
