+++ scripts/review/testdata/review-eval-split-equivalence/ledger.jsonl
@@ -1,0 +1,1 @@
+{"schema_version":1,"kind":"full","executed_at":"2026-08-28T02:13:20Z","status":"complete","verdict":"AMBER","comparability_key":"76c7cde728e17b7f758bcfa384fa64e5683f389d39ec058a9d551716dffe752f","contract_digest":"8b6479d106737fedea402269835428a5d29452acbc41c86d09bffee70fd2f9fb","inputs":{"skill_digest":"1afd9abdbe073cb23c29f635b5d9674ef927d319395e9f5d8672f7919454516a","skill_ref":"installed","finder_argv_digest":"40099c93f78800efefbc46f57772ae2cbadc86b378bdd22f474945abfd7b450c","orchestrator_digest":"2d7239cf858089e681eaa0942b901fb3961a1726f37b56e2960a878e69835892","claude_cli":"2.1.247 (Claude Code)","codex_cli":"codex-cli 0.150.1","host":"Mac"},"conditions":{"pipeline":{"model":"claude-opus-5","effort":"high","finder":"gpt-5.6-sol@high","draws":2,"recall":{"matched":30,"opportunities":68,"rate":0.441},"p1":{"matched":10,"opportunities":24,"rate":0.417},"novel_real":19,"wrong_claims":9,"usd":40.12,"seconds":18485,"per_defect":{"3827636772":[1,0],"3827647420":[0,0],"3827655676":[1,0],"3827655684":[1,0],"3827663924":[1,0],"3828303385":[1,1],"3828304403":[0,0],"3828313971":[0,0],"3828313974":[1,1],"3828328194":[1,1],"3829654081":[1,1],"3829659544":[1,1],"3829659546":[0,0],"3829659550":[0,1],"3829986398":[0,0],"3829986403":[1,0],"3830003055":[1,0],"3830042123":[0,1],"3830042128":[0,0],"3830259678":[0,0],"3830366234":[1,1],"3830366243":[1,1],"3830366247":[1,1],"3830366547":[0,0],"3830366556":[1,1],"3830371307":[0,0],"3830371315":[0,0],"3830371325":[1,1],"3830371329":[0,0],"3830394205":[0,0],"3830416461":[1,1],"3830423746":[0,0],"3830513980":[0,0],"3830695396":[0,0]}},"replay":{"model":"claude-opus-5","effort":"high","finder":"gpt-5.6-sol@high","draws":2,"recall":{"matched":21,"opportunities":44,"rate":0.477},"p1":{"matched":6,"opportunities":12,"rate":0.5},"novel_real":18,"wrong_claims":3,"usd":19.89,"seconds":3068,"per_defect":{"3829654081":[1,0],"3829659544":[1,1],"3829659546":[1,1],"3829659550":[0,0],"3829986398":[0,0],"3829986403":[0,0],"3830003055":[1,0],"3830042123":[1,1],"3830042128":[0,0],"3830259678":[0,0],"3830366234":[1,1],"3830366243":[1,1],"3830366247":[1,0],"3830366547":[1,0],"3830366556":[1,0],"3830371307":[1,0],"3830371315":[1,0],"3830371325":[1,1],"3830371329":[0,0],"3830394205":[0,0],"3830416461":[1,1],"3830695396":[0,0]}},"control":{"model":"claude-opus-5","effort":"high","draws":1,"recall":{"matched":11,"opportunities":34,"rate":0.324},"p1":{"matched":4,"opportunities":12,"rate":0.333},"novel_real":12,"wrong_claims":2,"usd":15.46,"seconds":2523,"per_defect":{"3827636772":[1],"3827647420":[0],"3827655676":[0],"3827655684":[1],"3827663924":[1],"3828303385":[1],"3828304403":[0],"3828313971":[0],"3828313974":[1],"3828328194":[1],"3829654081":[0],"3829659544":[1],"3829659546":[0],"3829659550":[0],"3829986398":[0],"3829986403":[0],"3830003055":[0],"3830042123":[1],"3830042128":[0],"3830259678":[0],"3830366234":[1],"3830366243":[0],"3830366247":[0],"3830366547":[0],"3830366556":[0],"3830371307":[1],"3830371315":[0],"3830371325":[1],"3830371329":[0],"3830394205":[0],"3830416461":[0],"3830423746":[0],"3830513980":[0],"3830695396":[0]}}},"judge_calibration":{"agreement":29,"total":40},"scoring_usd":32.54,"vs_baseline":null,"detail_dir":"docs/evals/review-skill-runs/2026-08-28-76c7cde7-full-1afd9abd","notes":""}
